separated JobProgress api and its implementation
